Restruturação do código
