+++ src/main/java/App.java
@@ -6,8 +6,8 @@ import main.java.gui.MainFrame;
 public class App {
 
 	public static void main(String[] args) {
-		MainFrame m = new MainFrame();
-		MainController c = new MainController(m);
+		
+		MainController c = new MainController();
 		c.init();
 	}
 }
