+++ src/main/java/gui/Popup_UploadFile.java
@@ -8,6 +8,7 @@ import javax.swing.JButton;
 import javax.swing.JFileChooser;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.SwingConstants;
 import javax.swing.WindowConstants;
@@ -17,22 +18,20 @@ import javax.swing.filechooser.FileSystemView;
 public class Popup_UploadFile {
 	private JFrame frame;
 	private JPanel panel;
-
+	private JButton import_button;
+	
 	public Popup_UploadFile() {
 		frame = new JFrame("Upload File");
 		frame.setLayout(new BorderLayout());
 		frame.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
-		addContents();
+		createCentralPanel();
+		frame.add(panel, BorderLayout.CENTER);
 		frame.setPreferredSize(new Dimension(300, 200));
 		frame.pack();
 		frame.setLocationRelativeTo(null);
 		frame.setVisible(true);
 	}
 
-	private void addContents() {
-		createCentralPanel();
-		frame.add(panel, BorderLayout.CENTER);
-	}
 	
 //TODO: Usar argumento Excel
 	private void createCentralPanel() { 
@@ -40,21 +39,23 @@ public class Popup_UploadFile {
 		panel.setLayout(new BorderLayout());
 	    JLabel label = new JLabel("Choose a file excel to import", SwingConstants.CENTER);
 	    panel.add(label, BorderLayout.CENTER );
-	    JButton b = new JButton("Click Here");
-	    panel.add(b, BorderLayout.SOUTH);
 	    
-		b.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				//TODO create method for listener
-				JFileChooser jfc = new JFileChooser(FileSystemView.getFileSystemView().getHomeDirectory());
-				int returnValue = jfc.showOpenDialog(null);
-				if (returnValue == JFileChooser.APPROVE_OPTION) {
-					File selectedFile = jfc.getSelectedFile();
-					System.out.println(selectedFile.getAbsolutePath());
-					frame.dispose();
-				}
-			}
-		});  
+	    import_button = new JButton("Click Here");
+	    panel.add(import_button, BorderLayout.SOUTH);    	    
 	}	
+	
+	public JButton getJButton() {
+		return import_button;
+	}
+	
+	public void close() {
+		frame.dispose();
+	}
+	
+	public void displayErrorMessage(String message) {
+		JOptionPane.showMessageDialog(frame,
+			    message,
+			    "Warning",
+			    JOptionPane.WARNING_MESSAGE);
+	}
 }
